Invert relationship between Kraken and mid-market rate
The mid-market rate is what we will use as a baseline to create
orders. Kraken is just one of many possible sources for the value of
the mid-market rate. It therefore makes sense to place `MidMarketRate`
at a higher level than the `kraken` module. The `MidMarketRate` struct
defines an interface which any mid-market rate source such as Kraken
needs to fulfil.
